Fix #382 For /target/tree/pharmacology/pages, trim SPARQL items Query..
Fixes https://github.com/openphacts/GLOBAL/issues/382

The default call for this LDA command was timing out.
Deleted the second half of the 'items' SPARQL query. That part of
the query consisted of a large UNION that seemed to collect triples
that were not actually used by the first part of the query that
actually limited the results returned.
Removing this appears to fix the problem.